Use the proper time-types for the job
The golang time.Time datatype records date and timezone information
that is utterly irrelevant for the subtitles.

The proper datatype to use is time.Duration, recording the duration
from the start of the video file that the subtitle shoud be
shown or hidden.

Potetial helper methods can return the duration of the subtitle
by subtracting the Start and End values